server/Python_Examples/bobbystable/requirements.txt
Flask
Flask_HTTPAuth
python-dotenv
signalwire_swaig
gunicorn
